+++ R/pkg.R
@@ -7,5 +7,4 @@
 NULL
 
 
-.current_model <- new.env()
-.current_model$udpipe_model <- structure(list(file = character(0), model = NULL), class = "udpipe_model")
+.loaded_models <- new.env()